+++ amber_lume/src/render/vulkan/buffer/typed/model_buffer.rs
@@ -28,7 +28,7 @@ impl ModelGpuData {
 
 pub fn create_model_buffer(
     buffer_factory: &ManagedBufferFactory,
-    capacity: usize,
+    capacity: u32,
 ) -> Result<PoolBuffer> {
     let item_size = size_of::<ModelGpuData>() as DeviceSize;
