+++ amber_lume/src/render/vulkan/buffer/typed/vertex_buffer.rs
@@ -45,7 +45,7 @@ impl VertexGpuData {
 }
 pub fn create_vertex_buffer(
     buffer_factory: &ManagedBufferFactory,
-    capacity: usize,
+    capacity: u32,
 ) -> Result<PoolBuffer> {
     let item_size = size_of::<VertexGpuData>() as DeviceSize;
